@@ -1,1 +1,2 @@
-from ._funcs import arg_or_default
+from ._image_funcs import add_corners, merge_alphas, pil_image_to_surface
+from ._funcs import arg_or_default, point_in_box
